search working very well now and purty
